Subject: Disposition of Submission -- SIDMA M129674RR
Von: <sidma@siam.org>
Date: So., 28. Feb. 2021 um 19:26 Uhr
To: <lf.kuehne@gmail.com>

Dear Lukas,

On behalf of the Editorial Board of SIAM J. Discrete Math (SIDMA), I am writing to you in regard to your submission "On the generation of rank 3 simple matroids with an application to Terao's freeness conjecture."

It is my pleasure to inform you that we are accepting your paper for publication in SIDMA. SIAM staff will be in contact within the next few days with details on where to send your TeX file and with further information on the publication process.

Thanks for allowing us to publish your paper, and keep us in mind in the future!

Best wishes,

James Oxley
Associate Editor
SIAM Journal on Discrete Mathematics